+++ src/main/java/io/electrum/moneytransfer/api/OrdersResource.java
@@ -21,8 +21,11 @@ import javax.ws.rs.core.UriInfo;
 import io.electrum.moneytransfer.model.ErrorDetail;
 import io.electrum.moneytransfer.model.MoneyTransferAuthRequest;
 import io.electrum.moneytransfer.model.MoneyTransferAuthResponse;
+import io.electrum.moneytransfer.model.MoneyTransferCancelOrderReversal;
 import io.electrum.moneytransfer.model.MoneyTransferConfirmation;
 import io.electrum.moneytransfer.model.MoneyTransferLookupResponse;
+import io.electrum.moneytransfer.model.MoneyTransferOrderCancelRequest;
+import io.electrum.moneytransfer.model.MoneyTransferOrderCancelResponse;
 import io.electrum.moneytransfer.model.MoneyTransferOrderStatusUpdateRequest;
 import io.electrum.moneytransfer.model.MoneyTransferRedeemRequest;
 import io.electrum.moneytransfer.model.MoneyTransferRedeemResponse;
@@ -99,6 +102,20 @@ public abstract class OrdersResource {
       public static final String ORDER_STATUS = "orderStatus";
       public static final int SUCCESS = 204;
       public static final String RELATIVE_PATH = CreateOrder.RELATIVE_PATH + "/status";
+      public static final String FULL_PATH = OrdersResource.PATH + RELATIVE_PATH;
+   }
+
+   public class CancelOrder {
+      public static final String CANCEL_ORDER = "cancelOrder";
+      public static final int SUCCESS = 200;
+      public static final String RELATIVE_PATH = "/cancel";
+      public static final String FULL_PATH = OrdersResource.PATH + RELATIVE_PATH;
+   }
+
+   public class CancelOrderReversal {
+      public static final String CANCEL_ORDER_REVERSAL = "cancelOrderReversal";
+      public static final int SUCCESS = 202;
+      public static final String RELATIVE_PATH = CancelOrder.RELATIVE_PATH + "/reversals";
       public static final String FULL_PATH = OrdersResource.PATH + RELATIVE_PATH;
    }
 
@@ -183,7 +200,6 @@ public abstract class OrdersResource {
    public final void createOrder(
          @ApiParam(value = "A message containing the data required to carry out the payment order, as well as "
                + "information about the point-of-sale from which the transaction originates.", required = true) @NotNull @Valid MoneyTransferAuthRequest body,
-
          @Context SecurityContext securityContext,
          @Context Request request,
          @Suspended AsyncResponse asyncResponse,
@@ -234,15 +250,10 @@ public abstract class OrdersResource {
          @ApiResponse(code = 504, message = "Gateway Timeout", response = ErrorDetail.class) })
    public final void lookupOrder(
          @ApiParam(value = "Reference used by the recipient to redeem the order. If both this value and remittanceRef are supplied then this takes precedence.") @QueryParam(LookupOrder.QueryParameters.ORDER_REDEEM_REF) String orderRedeemRef,
-
          @ApiParam(value = "The assigned merchant identifier. Also known as card acceptor id.") @QueryParam(LookupOrder.QueryParameters.MERCHANT_ID) String merchantId,
-
          @ApiParam(value = "Identifies the institution from which the transaction originates. Value to be assigned by Electrum.") @QueryParam(LookupOrder.QueryParameters.ORIGINATOR_INST_ID) String originatorInstId,
-
          @ApiParam(value = "Identifies the service provider to whom this request must be directed.", required = true) @QueryParam(LookupOrder.QueryParameters.RECEIVER_ID) @NotNull String receiverId,
-
          @ApiParam(value = "Reference used by the recipient to refer to the order when no redemption is to occur (e.g. direct account deposits). If both this value and orderRedeemRef are supplied then orderRedeemRef takes precedence.") @QueryParam(LookupOrder.QueryParameters.REMITTANCE_REF) String remittanceRef,
-
          @Context SecurityContext securityContext,
          @Context Request request,
          @Suspended AsyncResponse asyncResponse,
@@ -355,8 +366,7 @@ public abstract class OrdersResource {
    @Consumes({ "application/json" })
    @ApiOperation(value = OrderStatus.ORDER_STATUS, notes = "Advises of a change in an order's status.", authorizations = {
          @Authorization(value = "httpBasic") }, tags = {})
-   @ApiResponses(value = {
-         @ApiResponse(code = OrderStatus.SUCCESS, message = "No Content"),
+   @ApiResponses(value = { @ApiResponse(code = OrderStatus.SUCCESS, message = "No Content"),
          @ApiResponse(code = 400, message = "Bad request", response = ErrorDetail.class),
          @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDetail.class),
          @ApiResponse(code = 501, message = "Not implemented", response = ErrorDetail.class),
@@ -372,5 +382,62 @@ public abstract class OrdersResource {
          @Context HttpServletRequest httpServletRequest) {
       getResourceImplementation()
             .updateOrderStatus(body, securityContext, request, httpHeaders, asyncResponse, uriInfo, httpServletRequest);
+   }
+
+   @POST
+   @Consumes({ "application/json" })
+   @Produces({ "application/json" })
+   @ApiOperation(value = CancelOrder.CANCEL_ORDER, notes = "Requests that an order be cancelled and that a sender be refunded. This requested is treated as an online message. If no response is received, the cancellation will be reversed.", response = MoneyTransferOrderCancelResponse.class, authorizations = {
+         @Authorization(value = "httpBasic") }, tags = {})
+   @ApiResponses(value = {
+         @ApiResponse(code = CancelOrder.SUCCESS, message = "Ok", response = MoneyTransferOrderCancelResponse.class),
+         @ApiResponse(code = 400, message = "Bad request", response = ErrorDetail.class),
+         @ApiResponse(code = 404, message = "Not Found", response = ErrorDetail.class),
+         @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDetail.class),
+         @ApiResponse(code = 501, message = "Not implemented", response = ErrorDetail.class),
+         @ApiResponse(code = 503, message = "Service Unavailable", response = ErrorDetail.class),
+         @ApiResponse(code = 504, message = "Gateway Timeout", response = ErrorDetail.class) })
+   public final void cancelOrder(
+         @ApiParam(value = "A message containing the data required to carry out the payment order, as well as information about the point-of-sale from which the transaction originates.", required = true) @NotNull @Valid MoneyTransferOrderCancelRequest body,
+         @Context SecurityContext securityContext,
+         @Context Request request,
+         @Suspended AsyncResponse asyncResponse,
+         @Context HttpHeaders httpHeaders,
+         @Context UriInfo uriInfo,
+         @Context HttpServletRequest httpServletRequest) {
+      getResourceImplementation()
+            .cancelOrder(body, securityContext, request, httpHeaders, asyncResponse, uriInfo, httpServletRequest);
+   }
+
+   @POST
+   @Consumes({ "application/json" })
+   @Produces({ "application/json" })
+   @ApiOperation(value = CancelOrderReversal.CANCEL_ORDER_REVERSAL, notes = "Requests that a cancellation be reversed. This is treated as a store-and-forward transaction. Upon receipt, the Electrum service will respond immediately to acknowledge receipt and place the message on a queue for guaranteed delivery to the service provider.", response = MoneyTransferCancelOrderReversal.class, authorizations = {
+         @Authorization(value = "httpBasic") }, tags = {})
+   @ApiResponses(value = {
+         @ApiResponse(code = CancelOrderReversal.SUCCESS, message = "Accepted", response = MoneyTransferCancelOrderReversal.class),
+         @ApiResponse(code = 200, message = "OK", response = MoneyTransferCancelOrderReversal.class),
+         @ApiResponse(code = 400, message = "Bad request", response = ErrorDetail.class),
+         @ApiResponse(code = 404, message = "Not Found", response = ErrorDetail.class),
+         @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDetail.class),
+         @ApiResponse(code = 501, message = "Not implemented", response = ErrorDetail.class),
+         @ApiResponse(code = 503, message = "Service Unavailable", response = ErrorDetail.class),
+         @ApiResponse(code = 504, message = "Gateway Timeout", response = ErrorDetail.class) })
+   public final void cancelOrderReversal(
+         @ApiParam(value = "A message containing the data required to carry out the payment order, as well as information about the point-of-sale from which the transaction originates.", required = true) @NotNull @Valid MoneyTransferCancelOrderReversal body,
+         @Context SecurityContext securityContext,
+         @Context Request request,
+         @Suspended AsyncResponse asyncResponse,
+         @Context HttpHeaders httpHeaders,
+         @Context UriInfo uriInfo,
+         @Context HttpServletRequest httpServletRequest) {
+      getResourceImplementation().cancelOrderReversal(
+            body,
+            securityContext,
+            request,
+            httpHeaders,
+            asyncResponse,
+            uriInfo,
+            httpServletRequest);
    }
 }